health-first-server/requirements.txt
Flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.1
email-validator==2.1.0.post1
bcrypt==4.1.2
PyJWT==2.8.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
marshmallow==3.20.2
phonenumbers==8.13.30
pytest==8.0.0
pytest-flask==1.3.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
python-json-logger==2.0.7 